+++ packages/node-core/src/indexer/dynamic-ds.service.ts
@@ -1,14 +1,12 @@
 // Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import {isEqual, unionWith} from 'lodash';
 import {getLogger} from '../logger';
 import {CacheMetadataModel} from './storeCache/cacheMetadata';
 
 const logger = getLogger('dynamic-ds');
 
 const METADATA_KEY = 'dynamicDatasources';
-const TEMP_DS_PREFIX = 'ds_';
 
 export interface DatasourceParams {
   templateName: string;
@@ -24,14 +22,15 @@ export interface IDynamicDsService<DS> {
 
 export abstract class DynamicDsService<DS> implements IDynamicDsService<DS> {
   private _metadata?: CacheMetadataModel;
-  private tempDsRecords: Record<string, string> = {};
-
   private _datasources?: DS[];
+  private _datasourceParams?: DatasourceParams[];
+
+  protected abstract getDatasource(params: DatasourceParams): Promise<DS>;
 
   async init(metadata: CacheMetadataModel): Promise<void> {
     this._metadata = metadata;
 
-    this._datasources = await this.loadDynamicDatasources();
+    await this.getDynamicDatasources(true);
   }
 
   get dynamicDatasources(): DS[] {
@@ -52,25 +51,26 @@ export abstract class DynamicDsService<DS> implements IDynamicDsService<DS> {
    * remove dynamic ds that is created after this height
    * @param targetHeight this height is exclusive
    */
+  // eslint-disable-next-line @typescript-eslint/require-await
   async resetDynamicDatasource(targetHeight: number): Promise<void> {
-    const dynamicDs = await this.getDynamicDatasourceParams();
-    if (dynamicDs.length !== 0) {
-      const filteredDs = dynamicDs.filter((ds) => ds.startBlock <= targetHeight);
-      const dsRecords = JSON.stringify(filteredDs);
-      this.metadata.set(METADATA_KEY, dsRecords);
+    if (this._datasourceParams && this._datasourceParams.length !== 0) {
+      const filteredDs = this._datasourceParams.filter((ds) => ds.startBlock <= targetHeight);
+      this.metadata.set(METADATA_KEY, filteredDs);
+      await this.loadDynamicDatasources(filteredDs);
     }
   }
 
   async createDynamicDatasource(params: DatasourceParams): Promise<DS> {
     try {
       const ds = await this.getDatasource(params);
-
-      await this.saveDynamicDatasourceParams(params);
+      this.metadata.setNewDynamicDatasource(params);
 
       logger.info(`Created new dynamic datasource from template: "${params.templateName}"`);
 
       if (!this._datasources) this._datasources = [];
+      if (!this._datasourceParams) this._datasourceParams = [];
       this._datasources.push(ds);
+      this._datasourceParams.push(params);
 
       return ds;
     } catch (e: any) {
@@ -79,67 +79,24 @@ export abstract class DynamicDsService<DS> implements IDynamicDsService<DS> {
     }
   }
 
+  // Not force only seems to be used for project changes
   async getDynamicDatasources(forceReload?: boolean): Promise<DS[]> {
     // Workers should not cache this result in order to keep in sync
     if (!this._datasources || forceReload) {
-      this._datasources = await this.loadDynamicDatasources();
+      const params = (await this.metadata.find(METADATA_KEY)) ?? [];
+
+      await this.loadDynamicDatasources(params);
     }
 
-    return this._datasources;
+    // loadDynamicDatasources ensures this is set
+    return this._datasources as DS[];
   }
 
-  private async loadDynamicDatasources(): Promise<DS[]> {
-    try {
-      const params = await this.getDynamicDatasourceParams();
+  private async loadDynamicDatasources(params: DatasourceParams[]): Promise<void> {
+    const dataSources = await Promise.all(params.map((params) => this.getDatasource(params)));
 
-      const dataSources = await Promise.all(params.map((params) => this.getDatasource(params)));
-
-      logger.info(`Loaded ${dataSources.length} dynamic datasources`);
-
-      return dataSources;
-    } catch (e: any) {
-      logger.error(`Unable to get dynamic datasources:\n${e.message}`);
-      process.exit(1);
-    }
+    logger.info(`Loaded ${dataSources.length} dynamic datasources`);
+    this._datasourceParams = params;
+    this._datasources = dataSources;
   }
-
-  deleteTempDsRecords(blockHeight: number): void {
-    // Main thread will not have tempDsRecords with workers
-    if (this.tempDsRecords) {
-      delete this.tempDsRecords[TEMP_DS_PREFIX + blockHeight];
-    }
-  }
-
-  private async getDynamicDatasourceParams(blockHeight?: number): Promise<DatasourceParams[]> {
-    const record = await this.metadata.find(METADATA_KEY);
-
-    let results: DatasourceParams[] = [];
-
-    const metaResults: DatasourceParams[] = JSON.parse(record ?? '[]');
-    if (metaResults.length) {
-      results = [...metaResults];
-    }
-
-    if (blockHeight !== undefined) {
-      const tempResults: DatasourceParams[] = JSON.parse(this.tempDsRecords?.[TEMP_DS_PREFIX + blockHeight] ?? '[]');
-      if (tempResults.length) {
-        results = unionWith(results, tempResults, isEqual);
-      }
-    }
-
-    return results;
-  }
-
-  private async saveDynamicDatasourceParams(dsParams: DatasourceParams): Promise<void> {
-    const existing = await this.getDynamicDatasourceParams(dsParams.startBlock);
-
-    const dsRecords = JSON.stringify([...existing, dsParams]);
-    this.metadata.set(METADATA_KEY, dsRecords);
-    this.tempDsRecords = {
-      ...this.tempDsRecords,
-      [TEMP_DS_PREFIX + dsParams.startBlock]: dsRecords,
-    };
-  }
-
-  protected abstract getDatasource(params: DatasourceParams): Promise<DS>;
 }