+++ packages/node-core/src/indexer/fetch.service.ts
@@ -348,7 +348,6 @@ export abstract class BaseFetchService<DS extends BaseDataSource, B extends IBlo
   }
 
   resetForNewDs(blockHeight: number): void {
-    this.dynamicDsService.deleteTempDsRecords(blockHeight);
     this.updateDictionary();
     this.blockDispatcher.flushQueue(blockHeight);
   }