+++ packages/node-core/src/indexer/project.service.ts
@@ -22,6 +22,7 @@ import {
 import {BlockHeightMap} from '../utils/blockHeightMap';
 import {BaseDsProcessorService} from './ds-processor.service';
 import {DynamicDsService} from './dynamic-ds.service';
+import {MetadataKeys} from './entities';
 import {PoiSyncService} from './poi';
 import {PoiService} from './poi/poi.service';
 import {StoreService} from './store.service';
@@ -108,8 +109,9 @@ export abstract class BaseProjectService<
 
       // Init metadata before rest of schema so we can determine the correct project version to create the schema
       await this.storeService.initCoreTables(this._schema);
-      await this.dynamicDsService.init(this.storeService.storeCache.metadata);
       await this.ensureMetadata();
+      // DynamicDsService is dependent on metadata so we need to ensure it exists first
+      await this.dynamicDsService.init(this.storeService.storeCache.metadata);
 
       /**
        * WARNING: The order of the following steps is very important.
@@ -198,7 +200,7 @@ export abstract class BaseProjectService<
 
     this.eventEmitter.emit(IndexerEvent.NetworkMetadata, this.apiService.networkMeta);
 
-    const keys = [
+    const keys: (keyof MetadataKeys)[] = [
       'lastProcessedHeight',
       'blockOffset',
       'indexerNodeVersion',
@@ -209,7 +211,8 @@ export abstract class BaseProjectService<
       'processedBlockCount',
       'lastFinalizedVerifiedHeight',
       'schemaMigrationCount',
-    ] as const;
+      'dynamicDatasources',
+    ];
 
     const existing = await metadata.findMany(keys);
 
@@ -255,6 +258,14 @@ export abstract class BaseProjectService<
     }
     if (!existing.startHeight) {
       metadata.set('startHeight', this.getStartBlockFromDataSources());
+    }
+
+    if (!existing.dynamicDatasources) {
+      metadata.set('dynamicDatasources', []);
+    } else if (typeof existing.dynamicDatasources === 'string') {
+      // Migration Step: In versions  < 4.7.2 dynamic datasources was stored as a string in a json field.
+      logger.info('Migrating dynamic datasources from string to object');
+      metadata.set('dynamicDatasources', JSON.parse(existing.dynamicDatasources));
     }
   }